+++ trinity/display/x11/core/impl/src/main/java/org/trinity/display/x11/core/impl/event/ButtonPressConversion.java
@@ -52,7 +52,7 @@ public class ButtonPressConversion implements XEventConversion {
 																							true);
 		this.xEventBus.post(button_press_event_t);
 
-		final int windowId = button_press_event_t.getEvent();
+		final int windowId = (int) button_press_event_t.getEvent();
 		final XWindow window = this.windowCache.getWindow(windowId);
 
 		final int buttonCode = button_press_event_t.getDetail();
