+++ trinity/display/x11/core/impl/src/main/java/org/trinity/display/x11/core/impl/event/CirculateNotifyConversion.java
@@ -40,7 +40,7 @@ public class CirculateNotifyConversion implements XEventConversion {
 																										true);
 		this.xEventBus.post(circulate_notify_event_t);
 
-		final int windowId = circulate_notify_event_t.getEvent();
+		final int windowId = (int) circulate_notify_event_t.getEvent();
 		final XWindow displayEventSource = this.xWindowCache.getWindow(windowId);
 
 		final DisplayEvent displayEvent = new StackingChangedNotifyEvent(displayEventSource);
